api/requirements.txt
tensorflow==2.16.1
flask
numpy
pillow